fix: wrong way to sanitize files names for file upload
